// apps/web/src/types/dao.ts
export interface Member {
  addr: string
  weight: number
  [k: string]: unknown
}